requirements.txt
numpy==2.0.2 #2.3.5
torch==2.9.1
accelerate==1.12.0
Pillow==11.3.0
transformers==4.57.5
diffusers==0.36.0
ultralytics==8.3.253
streamlit==1.52.2
tqdm==4.67.1
huggingface-hub==0.34.0
imageio==2.37.2
imageio_ffmpeg==0.6.0
# Создано с помощью pipreqs
